refactor: Split monolithic 'owner' package into domain-specific directories
This commit reorganizes the directory structure to improve maintainability and separation of concerns.

The classes previously located in the single 'owner' directory have been moved into new, domain-specific packages:
- `owner/`: For all owner-related classes.
- `pet/`: For all pet-related classes.
- `visits/`: For all visit-related classes.
 Changes to be committed:
