+++ src/main/java/org/springframework/samples/petclinic/owner/Owner.java
@@ -20,6 +20,8 @@ import java.util.List;
 
 import org.springframework.core.style.ToStringCreator;
 import org.springframework.samples.petclinic.model.Person;
+import org.springframework.samples.petclinic.pet.Pet;
+import org.springframework.samples.petclinic.visit.Visit;
 import org.springframework.util.Assert;
 
 import jakarta.persistence.CascadeType;
