+++ src/main/java/org/springframework/samples/petclinic/pet/PetTypeRepository.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.samples.petclinic.owner;
+package org.springframework.samples.petclinic.pet;
 
 import java.util.List;
 import java.util.Optional;
