+++ src/test/java/org/springframework/samples/petclinic/owner/PetTypeFormatterTests.java
@@ -32,6 +32,9 @@ import org.junit.jupiter.api.condition.DisabledInNativeImage;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.samples.petclinic.pet.PetType;
+import org.springframework.samples.petclinic.pet.PetTypeFormatter;
+import org.springframework.samples.petclinic.pet.PetTypeRepository;
 
 /**
  * Test class for {@link PetTypeFormatter}
