+++ src/test/java/org/springframework/samples/petclinic/owner/PetValidatorTests.java
@@ -22,6 +22,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledInNativeImage;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.samples.petclinic.pet.Pet;
+import org.springframework.samples.petclinic.pet.PetType;
+import org.springframework.samples.petclinic.pet.PetValidator;
 import org.springframework.validation.Errors;
 import org.springframework.validation.MapBindingResult;
